--- a/utils/loadCommands.js
+++ b/utils/loadCommands.js
@@ -7,7 +7,7 @@ const loadCommands = (client, baseDir) => new Promise(async (resolve, reject) =>
     await fs.ensureDirAsync(dir).catch(err => client.funcs.log(err, "error"));
     const files = await client.funcs.getFileListing(client, baseDir, "commands").catch(err => client.emit("error", client.funcs.newError(err)));
     const fn = files.map(f => client.funcs.loadSingleCommand(client, `${f.name}`, false, `${f.path}${path.sep}${f.base}`));
-    await Promise.all(fn);
+    await Promise.all(fn).catch(e => client.funcs.log(e, "error"));
     resolve();
   } catch (e) {
     if (e.code === "MODULE_NOT_FOUND") {
@@ -27,9 +27,9 @@ const loadCommands = (client, baseDir) => new Promise(async (resolve, reject) =>
 module.exports = async (client) => {
   client.commands.clear();
   client.aliases.clear();
-  await loadCommands(client, client.coreBaseDir).catch(err => client.emit("error", client.funcs.newError(err)));
+  await loadCommands(client, client.clientBaseDir).catch(err => client.emit("error", client.funcs.newError(err)));
   if (client.coreBaseDir !== client.clientBaseDir) {
-    await loadCommands(client, client.clientBaseDir).catch(err => client.emit("error", client.funcs.newError(err)));
+    await loadCommands(client, client.coreBaseDir).catch(err => client.emit("error", client.funcs.newError(err)));
   }
   client.funcs.log(`Loaded ${client.commands.size} commands, with ${client.aliases.size} aliases.`);
 };