--- a/utils/loadMessageMonitors.js
+++ b/utils/loadMessageMonitors.js
@@ -6,11 +6,17 @@ const loadMessageMonitors = (client, baseDir) => new Promise(async (resolve, rej
   await fs.ensureDirAsync(dir).catch(err => client.emit("error", client.funcs.newError(err)));
   const files = await client.funcs.getFileListing(client, baseDir, "monitors").catch(err => client.emit("error", client.funcs.newError(err)));
   try {
-    files.forEach((f) => {
-      const props = require(`${f.path}${path.sep}${f.base}`);
-      if (props.init) props.init(client);
-      client.messageMonitors.set(f.name, props);
-    });
+    const fn = files.map(f => new Promise((res, rej) => {
+      try {
+        const props = require(`${f.path}${path.sep}${f.base}`);
+        if (props.init) props.init(client);
+        client.messageMonitors.set(f.name, props);
+        res();
+      } catch (e) {
+        rej(e);
+      }
+    }));
+    await Promise.all(fn);
     resolve();
   } catch (e) {
     if (e.code === "MODULE_NOT_FOUND") {