--- a/utils/loadMessageMonitors.js
+++ b/utils/loadMessageMonitors.js
@@ -11,7 +11,7 @@ const loadMessageMonitors = (client, baseDir) => new Promise(async (resolve, rej
       const props = require(`${f.path}${path.sep}${f.base}`);
       if (props.init) props.init(client);
       client.messageMonitors.set(f.name, props);
-      res();
+      res(delete require.cache[require.resolve(`${f.path}${path.sep}${f.base}`)]);
     }));
     await Promise.all(fn).catch(e => client.funcs.log(e, "error"));
     resolve();