--- a/utils/loadProviders.js
+++ b/utils/loadProviders.js
@@ -6,11 +6,17 @@ const loadProviders = (client, baseDir) => new Promise(async (resolve, reject) =
   await fs.ensureDirAsync(dir).catch(err => client.emit("error", client.funcs.newError(err)));
   const files = await client.funcs.getFileListing(client, baseDir, "providers").catch(err => client.emit("error", client.funcs.newError(err)));
   try {
-    files.forEach((f) => {
-      const props = require(`${f.path}${path.sep}${f.base}`);
-      if (props.init) props.init(client);
-      client.providers.set(f.name, props);
-    });
+    const fn = files.map(f => new Promise((res, rej) => {
+      try {
+        const props = require(`${f.path}${path.sep}${f.base}`);
+        if (props.init) props.init(client);
+        client.providers.set(f.name, props);
+        res();
+      } catch (e) {
+        rej(e);
+      }
+    }));
+    await Promise.all(fn);
     resolve();
   } catch (e) {
     if (e.code === "MODULE_NOT_FOUND") {