… functions/loadCommandInhibitors.js
… functions/loadCommands.js
… functions/loadEvents.js
… functions/loadFunctions.js
… functions/loadMessageMonitors.js
… functions/loadProviders.js
… functions/parseTags.js
… package.json