### Proposed Semver Increment Bump: [MAJOR]

Loading times, tested in two bots:
108 commands, (~10.200loc) 35s => 6s (5.8x faster)
78 commands, (~6.400loc) 24s => 5s (4.8x faster)

### Changes Proposed in this Pull Request (List new items in CHANGELOG.MD)

- Faster loading times
- Cache optimization

### Fixes:

- TypeError from `awaitMessage`, as in #158