@kyranet 2017-03-20
I added a last change, deleting the file from the cache after loading the pieces. By that, I could reduce 5MB worth of data from the cache with this.

On my point of view, I don't think we need the same data cached in both `require.cache` and `client.CollectionOfPieces`. I also thought this change *may* break the reload command, but I have just tested and it doesn't return an error.