@kyranet 2017-03-21
According to the reload command. 

- Reload command: **OK!**
- Reload monitor: **OK!**
- Reload inhibitor: **OK!**
- Reload function: **OK!**

Memory usage seems to be better, dropped 8MB with all the previous PRs (now using 41MB vs 49MB). As well as the loading time is now faster since we removed a bunch of useless functions. ¯\\\_(ツ)\_/¯