@kyranet 2017-03-21
Lastest commit fixes the issue #158.